// src/app/api/swagger/models/api-participant-offer-dto.ts
export interface ApiParticipantOfferDto {
  Category?: ApiCategoryDto | null;
  CharacteristicFileGuid?: string;
  CharacteristicFromUi?: boolean | null;
  CharacteristicValues?: Array<ApiCharacteristicValuesDto> | null;
  ClassificatorCode?: string | null;
  ClassificatorDescription?: string | null;
  Country?: ApiCountryDto | null;
  CreatedMethod?: ApiOfferCreatedMethodEnum;
  CurrencyCode?: string | null;
  DeliveryCost?: number | null;
  DeliveryPlace?: string | null;
  DeliveryRegionInfos?: Array<ApiDeliveryRegionInfoDto> | null;
  DeliveryRegions?: string | null;
  DuringDays?: number | null;
  ExclusiveToTrade?: boolean | null;
  ExternalSystemIdentifier?: string | null;
  ExternalSystemIdentifierNumeric?: number | null;
  Id?: number;
  IsPublicOffer?: boolean;
  ItemInStock?: boolean | null;
  KsrCode?: string | null;
  LotItemId?: number | null;
  Manufacturer?: string | null;
  ManufacturerCountryRegistryNumber?: string | null;
  ManufacturerCountryRegistryType?: ApiManufacturerCountryRegistryTypeEnum | null;
  ManufacturerCountryScore?: number | null;
  MinQuantity?: number | null;
  NeverExpired?: boolean;
  OfferState?: ApiParticipantOfferStateEnum | null;
  OfferTerms?: string | null;
  OfferType?: ApiParticipantOfferTypeEnum;
  OkeiCode?: string | null;
  OkeiDescription?: string | null;
  Okpd2Codes?: Array<ApiOkpd2MarketDto> | null;
  ParticipantOfferDocuments?: Array<ApiParticipantOfferDocumentDto> | null;
  ParticipantOfferSource?: ApiParticipantOfferSourceEnum;
  ParticipantOfferSpecific?: ApiParticipantOfferSpecificFlagEnum | null;
  PossibleDeliveryBySeller?: boolean;
  PossibleSelfDelivery?: boolean | null;
  PossibleSendByTransportCompany?: boolean | null;
  PriceListId?: number | null;
  PriceRegionDetails?: Array<ApiPriceRegionDetailDto> | null;
  PrintFormGuid?: string | null;
  ProductDescription?: string | null;
  ProductName?: string | null;
  PublishDateStart?: string | null;
  Quantity?: number | null;
  RecipientOrganizationInn?: string | null;
  RecipientOrganizationKpp?: string | null;
  RecipientOrganizationName?: string | null;
  RegionalProductSignChecked?: boolean;
  Regions?: Array<ApiRegionKladrDto> | null;
  RequestDateEnd?: string | null;
  RequestDateStart?: string | null;
  SelfDeliveryPlace?: string | null;
  Signature?: string | null;
  TradeId?: number | null;
  TruType?: ApiTruTypeEnum2 | null;
  UnitPrice?: number | null;
  UnitPriceInCurrency?: number | null;
  UsedClassificatorType?: ApiUsedClassificatorTypeEnum;
  VatRate?: number | null;
  VatRateState?: ApiVatRateStateEnum;
}